+++ poc/classes/AuxSymbolTable.py
@@ -10,7 +10,9 @@ from poc.classes.AuxInterpretation import AuxInterpretation
 from poc.classes.NamedVariableDeclaration import NamedVariableDeclaration
 from anytree import AnyNode, Resolver
 import poc.fplerror
-import copy
+from poc.classes.Signature import Signature
+from poc.classes.AuxBits import AuxBits
+from poc.classes.AuxContext import AuxContext
 
 
 class AuxSymbolTable:
@@ -23,6 +25,7 @@ class AuxSymbolTable:
     classDeclaration = "class"
     classProperty = "property"
     conjectures = "conjectures"
+    conjecture = "conjecture"
     functionalTerm = "functionalTerm"
     functionalTermImage = "image"
     functionalTerms = "functionalTerms"
@@ -47,7 +50,6 @@ class AuxSymbolTable:
     theory = "theory"
     theoryName = "theory_name"
     type = "type"
-    usedNamespace = "usedNamespace"
     uses = "uses"
     var = "var"
     variables = "variables"
@@ -90,7 +92,7 @@ class AuxSymbolTable:
         return r.get(parent, outline)
 
     @staticmethod
-    def _register_global_reference(theory_node: AnyNode, global_id: str, node: AnyNode, errors: list):
+    def _register_global_reference(theory_node: AnyNode, gid: str, uid: str, node: AnyNode, errors: list):
         root = theory_node.parent
         global_references_node = AuxSymbolTable.get_child_by_outline(root, AuxSymbolTable.globalLookup)
         # there can be nodes with the same global id in the symbol table. This can happen if the
@@ -98,15 +100,15 @@ class AuxSymbolTable:
         try:
             # There cannot be nodes with the same global id in the symbol table. If this happens, we have some bug
             # in the proceeding steps.
-            node = r.get(global_references_node, global_id)
-            raise AssertionError("Name conflict was not discovered for " + global_id + " and " + str(node.info))
+            node = r.get(global_references_node, gid)
+            raise AssertionError("Name conflict was not discovered for " + gid + " and " + str(node.info))
         except anytree.resolver.ChildResolverError:
             at_least_one_error = False
-            significant_name = global_id.split(".")[-1]
+            significant_name = gid.split(".")[-1]
             # It can happen that the significant name is the same as some part other of the gid (for instance,
             # an FPL class could be named the same as the namespace or the same as its property). In this case, we
             # have another error.
-            names = global_id.split(".")
+            names = gid.split(".")
             for name in names[:-1]:
                 if significant_name == name:
                     # Note that the constructor's name of a class is an exception from this rule:
@@ -115,12 +117,30 @@ class AuxSymbolTable:
                         # ok
                         pass
                     else:
-                        errors.append(poc.fplerror.FplMalformedGlobalId(node.info, global_id))
+                        errors.append(poc.fplerror.FplMalformedGlobalId(node.info, gid))
                         at_least_one_error = True
 
             if not at_least_one_error:
                 # if there were no errors, we can register the new node
-                AnyNode(parent=global_references_node, outline=node.outline, gid=global_id, node=node)
+                global_node = AnyNode(parent=global_references_node, outline=node.outline, gid=gid, uid=uid, node=node)
+                # set the global node's  type pattern
+                if node.outline == AuxSymbolTable.classConstructor:
+                    global_node.type_pattern = AuxBits.isObject
+                elif node.outline == AuxSymbolTable.classDeclaration:
+                    global_node.type_pattern = AuxBits.isClass
+                    global_node.inherits = AnyNode()
+                elif node.outline == AuxSymbolTable.property:
+                    global_node.type_pattern = node.type_pattern
+                elif node.outline in [AuxSymbolTable.theoremLikeStmt, AuxSymbolTable.conjecture,
+                                      AuxSymbolTable.predicateDeclaration, AuxSymbolTable.inferenceRule,
+                                      AuxSymbolTable.axiom]:
+                    global_node.type_pattern = AuxBits.isPredicate
+                elif node.outline == AuxSymbolTable.functionalTerm:
+                    global_node.type_pattern = AuxBits.isFunctionalTerm
+                else:
+                    raise NotImplementedError(node.outline)
+                return global_node
+        return None
 
     @staticmethod
     def add_usage_to_theory(theory_node: AnyNode, parsing_info: AuxInterpretation):
@@ -132,9 +152,9 @@ class AuxSymbolTable:
             parsing_info.all_errors().append(
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
-            node = AnyNode(parent=uses_node, outline=AuxSymbolTable.usedNamespace, id=identifier,
+            node = AnyNode(parent=uses_node, id=identifier, modifier=parsing_info.modifier,
                            info=parsing_info.get_ast_info())
-        return node
+            return node
 
     @staticmethod
     def add_inference_rule_to_theory(theory_node: AnyNode, parsing_info: AuxInterpretation):
@@ -148,9 +168,6 @@ class AuxSymbolTable:
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=inference_rules_node, outline=AuxSymbolTable.inferenceRule, id=identifier,
                            info=parsing_info.get_ast_info())
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
         return node
@@ -166,14 +183,15 @@ class AuxSymbolTable:
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=classes_node, outline=AuxSymbolTable.classDeclaration, id=identifier,
-                           info=parsing_info.get_ast_info(), type=None)
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
+                           info=parsing_info.get_ast_info())
             # add properties and constructors subnodes to this node
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
             AnyNode(parent=node, outline=AuxSymbolTable.classConstructors)
             AnyNode(parent=node, outline=AuxSymbolTable.properties)
+            # register the global reference for the class
+            theory_node = classes_node.parent
+            gid = '.'.join([theory_node.namespace, node.id])
+            AuxSymbolTable._register_global_reference(theory_node, gid, node.id, node, parsing_info.all_errors())
         return node
 
     @staticmethod
@@ -187,10 +205,7 @@ class AuxSymbolTable:
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=predicates_node, outline=AuxSymbolTable.predicateDeclaration, id=identifier,
-                           info=parsing_info.get_ast_info(), overload_id=None)
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
+                           info=parsing_info.get_ast_info())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
         return node
@@ -206,10 +221,7 @@ class AuxSymbolTable:
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=functional_terms_node, outline=AuxSymbolTable.functionalTerm, id=identifier,
-                           info=parsing_info.get_ast_info(), overload_id=None)
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
+                           info=parsing_info.get_ast_info())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.functionalTermImage)
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
@@ -234,28 +246,27 @@ class AuxSymbolTable:
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=axiom_nodes, outline=AuxSymbolTable.axiom, id=identifier,
                            info=parsing_info.get_ast_info())
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
         return node
 
     @staticmethod
-    def add_theorem_like_stmt(theory_node: AnyNode, parsing_info: AuxInterpretation, statement_type: str):
-        theorem_like_nodes = AuxSymbolTable.get_child_by_outline(theory_node, AuxSymbolTable.theoremLikeStmts)
+    def add_theorem_like_stmt_or_conj(theory_node: AnyNode, parsing_info: AuxInterpretation, statement_type: str):
+        parent = AuxSymbolTable.get_child_by_outline(theory_node, AuxSymbolTable.theoremLikeStmts)
+        outline = AuxSymbolTable.theoremLikeStmt
+        # correct the parent and outline, if the statement is a conjecture
+        if statement_type == AuxContext.theoremLikeStmtConj:
+            parent = AuxSymbolTable.get_child_by_outline(theory_node, AuxSymbolTable.conjectures)
+            outline = AuxSymbolTable.conjecture
         identifier = parsing_info.id
         r = Resolver(AuxSymbolTable.ids)
         try:
-            node = r.get(theorem_like_nodes, identifier)
+            node = r.get(parent, identifier)
             parsing_info.all_errors().append(
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
-            node = AnyNode(parent=theorem_like_nodes, outline=AuxSymbolTable.theoremLikeStmt, id=identifier,
+            node = AnyNode(parent=parent, outline=outline, id=identifier,
                            info=parsing_info.get_ast_info(), statement_type=statement_type)
-            # register the node globally
-            gid = '.'.join([theory_node.namespace, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
         return node
@@ -270,18 +281,14 @@ class AuxSymbolTable:
 
         r = Resolver(AuxSymbolTable.outline)
         constructors_node = r.get(class_node, AuxSymbolTable.classConstructors)
+        r = Resolver(AuxSymbolTable.ids)
         try:
-            r = Resolver(AuxSymbolTable.ids)
             node = r.get(constructors_node, identifier)
             parsing_info.all_errors().append(
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=constructors_node, outline=AuxSymbolTable.classConstructor, id=identifier,
                            info=parsing_info.get_ast_info())
-            # register the node globally
-            theory_node = class_node.parent.parent
-            gid = '.'.join([theory_node.namespace, class_node.id, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
             # add specific subnodes
             AnyNode(parent=node, outline=AuxSymbolTable.variables)
         return node
@@ -308,13 +315,8 @@ class AuxSymbolTable:
                 poc.fplerror.FplIdentifierAlreadyDeclared(parsing_info.id, parsing_info.get_ast_info(), node.info))
         except anytree.resolver.ChildResolverError:
             node = AnyNode(parent=properties_node, outline=AuxSymbolTable.property, id=identifier,
-                           info=parsing_info.get_ast_info(), is_mandatory=is_mandatory, type=property_type.id,
-                           overload_id=None)
+                           info=parsing_info.get_ast_info(), is_mandatory=is_mandatory)
 
-            # register the node globally
-            theory_node = parent.parent.parent
-            gid = '.'.join([theory_node.namespace, parent.id, node.id])
-            AuxSymbolTable._register_global_reference(theory_node, gid, node, parsing_info.all_errors())
             # in case of functional term nodes, add image as a subnode
             if property_type.id in ["func", "function"]:
                 AnyNode(parent=node, outline=AuxSymbolTable.functionalTermImage)
@@ -330,9 +332,8 @@ class AuxSymbolTable:
     @staticmethod
     def add_param_to_image_node(image_node: AnyNode, parsing_info: AuxInterpretation):
         identifier = parsing_info.id
-        param_no = str(len(image_node.children) + 1)
         AnyNode(parent=image_node, info=parsing_info.get_ast_info(), id=identifier,
-                outline=AuxSymbolTable.param + param_no)
+                outline=AuxSymbolTable.param)
 
     @staticmethod
     def clone_tree(source: AnyNode):
@@ -366,11 +367,8 @@ class AuxSymbolTable:
                                id=var.id,
                                outline=AuxSymbolTable.var,
                                type=named_var_decl.var_type.generalType.id,
-                               type_modifier=named_var_decl.var_type.generalType.callModifier,
-                               type_inBuilt=named_var_decl.var_type.generalType.inBuilt,
-                               type_generic=named_var_decl.var_type.generalType.generic,
-                               type_isIndex=named_var_decl.var_type.generalType.isIndex,
-                               type_isSyntaxExtension=named_var_decl.var_type.generalType.isSyntaxExtension,
+                               type_pattern=named_var_decl.var_type.generalType.pattern_int,
+                               type_mod=named_var_decl.var_type.generalType.mod,
                                info=var.get_ast_info(),
                                is_signature=is_signature
                                )
@@ -378,3 +376,39 @@ class AuxSymbolTable:
                 for next_var_decl in reversed(named_var_decl.var_type.paramTuple.tuple):
                     AuxSymbolTable.add_vars_to_nodes(var_node, next_var_decl, is_signature)
             var_node.parent = parent
+
+    @staticmethod
+    def set_global_id(building_block_node: AnyNode, signature: Signature):
+        """
+        Sets the (cross-namespace) global id for a given building block of a the FPL theory. This is unique
+        globally but reflects overriding in FPL. This is accomplished by using different derivations
+        signatures with the same PredicateIdentifier of the building block.
+        :param building_block_node: a building block node in the symbol table whose id is some PredicateIdentifier
+        :param signature: A Signature instance that was created by the FPLInterpreter during the parsing process.
+        :return: None
+        """
+        parent = building_block_node.parent
+        r = Resolver(AuxSymbolTable.ids)
+        override_name = building_block_node.id + signature.to_signature_string()
+        try:
+            other = r.get(parent, override_name)
+            # the override id already exists
+            signature.all_errors().append(
+                poc.fplerror.FplIdentifierAlreadyDeclared(building_block_node.id, building_block_node.info, other.info))
+        except anytree.resolver.ChildResolverError:
+            uid = building_block_node.id
+            building_block_node.id = override_name
+            theory_node = None
+            if building_block_node.outline in [AuxSymbolTable.classConstructor, AuxSymbolTable.property]:
+                theory_node = parent.parent.parent.parent
+                gid = '.'.join([theory_node.namespace, parent.parent.id, building_block_node.id])
+            else:
+                theory_node = parent.parent
+                gid = '.'.join([theory_node.namespace, building_block_node.id])
+            AuxSymbolTable._register_global_reference(theory_node, gid, uid, building_block_node,
+                                                      signature.all_errors())
+
+            # add signature to node's info because we will need it whenever the FPL interpreter
+            # identifies a reference to the node in the FPL code ("call") and will have to check if this references
+            # uses a compatible parameter specifications to use the "call".
+            building_block_node.info.signature = signature