+++ poc/demo.py
@@ -5,7 +5,7 @@ u = fplutil.Utils()
 fpl_parser = u.get_parser("../grammar/fpl_tatsu_format.ebnf")
 u.add_distinct_duration("creating tatsu parser")
 
-theoryCommon = u.get_file_content("theories/Commons.fpl")
+theoryCommon = fplutil.Utils.get_file_content("theories/Commons.fpl")
 interpreterCommon = FplInterpreter(fpl_parser)
 interpreterCommon.syntax_transform("common", theoryCommon)
 u.add_distinct_duration("interpreting common")
@@ -13,7 +13,7 @@ print(interpreterCommon.minified())
 if interpreterCommon.has_errors():
     interpreterCommon.print_errors()
 
-theoryNat = u.get_file_content("theories/ArithmeticsNat.fpl")
+theoryNat = fplutil.Utils.get_file_content("theories/ArithmeticsNat.fpl")
 interpreterNat = FplInterpreter(fpl_parser)
 interpreterNat.syntax_transform("nat", theoryCommon)
 u.add_distinct_duration("interpreting nat")
@@ -21,7 +21,7 @@ print(interpreterNat.minified())
 if interpreterNat.has_errors():
     interpreterNat.print_errors()
 
-theorySet = u.get_file_content("theories/Set.fpl")
+theorySet = fplutil.Utils.get_file_content("theories/Set.fpl")
 interpreterSet = FplInterpreter(fpl_parser)
 interpreterSet.syntax_transform("settheory", theorySet)
 u.add_distinct_duration("interpreting set")
@@ -29,7 +29,7 @@ print(interpreterSet.minified())
 if interpreterSet.has_errors():
     interpreterSet.print_errors()
 
-theoryAlgebra = u.get_file_content("theories/Algebra.fpl")
+theoryAlgebra = fplutil.Utils.get_file_content("theories/Algebra.fpl")
 interpreterAlgebra = FplInterpreter(fpl_parser)
 interpreterAlgebra.syntax_transform("algebra", theoryAlgebra)
 u.add_distinct_duration("interpreting algebra")
@@ -37,7 +37,7 @@ print(interpreterAlgebra.minified())
 if interpreterAlgebra.has_errors():
     interpreterAlgebra.print_errors()
 
-theoryGeometry = u.get_file_content("theories/Geometry.fpl")
+theoryGeometry = fplutil.Utils.get_file_content("theories/Geometry.fpl")
 interpreterGeometry = FplInterpreter(fpl_parser)
 interpreterGeometry.syntax_transform("geometry", theoryGeometry)
 u.add_distinct_duration("interpreting geometry")
@@ -45,7 +45,7 @@ print(interpreterGeometry.minified())
 if interpreterGeometry.has_errors():
     interpreterGeometry.print_errors()
 
-theoryExample25 = u.get_file_content("theories/Example4-7.fpl")
+theoryExample25 = fplutil.Utils.get_file_content("theories/Example4-7.fpl")
 interpreterExample25 = FplInterpreter(fpl_parser)
 interpreterExample25.syntax_transform("example25", theoryExample25)
 u.add_distinct_duration("interpreting example25")
@@ -53,7 +53,7 @@ print(interpreterExample25.minified())
 if interpreterExample25.has_errors():
     interpreterExample25.print_errors()
 
-theoryLinalg = u.get_file_content("theories/Linalg.fpl")
+theoryLinalg = fplutil.Utils.get_file_content("theories/Linalg.fpl")
 interpreterLinalg = FplInterpreter(fpl_parser)
 interpreterLinalg.syntax_transform("linalg", theoryLinalg)
 u.add_distinct_duration("interpreting linalg")
@@ -61,7 +61,7 @@ print(interpreterLinalg.minified())
 if interpreterLinalg.has_errors():
     interpreterLinalg.print_errors()
 
-theoryComStruct = u.get_file_content("theories/CommonsStructures.fpl")
+theoryComStruct = fplutil.Utils.get_file_content("theories/CommonsStructures.fpl")
 interpreterComStruct = FplInterpreter(fpl_parser)
 interpreterComStruct.syntax_transform("comstruct", theoryComStruct)
 u.add_distinct_duration("interpreting comstruct")