+++ poc/fplsourceanalyser.py
@@ -341,7 +341,8 @@ class FPLSourceAnalyser(object):
         rule = parsing_info.rule_name()
         func = self.switcher.get(rule, lambda: "Invalid rule")
         if self.i.verbose:
-            print("switched " + (str(func)).split(" ")[1] + " for " + rule)
+            print("switched " + (str(func)).split(" ")[1] + " for " + str(parsing_info) + " in context " + str(
+                self.i.context.get_context()))
         # call the function depending on the rule in the switcher and modify its interpretation and returns its value
         return func(self.i, parsing_info)